+++ apps/linows/src-tauri/src/autostart.rs
@@ -1,52 +1,39 @@
-use std::path::PathBuf;
-
-fn autostart_dir() -> PathBuf {
-    let config = std::env::var("XDG_CONFIG_HOME").unwrap_or_else(|_| {
-        let home = std::env::var("HOME").unwrap_or_else(|_| ".".to_string());
-        format!("{home}/.config")
-    });
-    PathBuf::from(config).join("autostart")
-}
-
-fn desktop_entry_path() -> PathBuf {
-    autostart_dir().join("look.desktop")
-}
-
-fn current_exe_path() -> String {
-    std::env::current_exe()
-        .map(|p| p.to_string_lossy().to_string())
-        .unwrap_or_else(|_| "lookapp".to_string())
-}
+//! Cross-platform autostart Tauri commands. Real per-OS implementations live in
+//! `platform::linux::autostart` and `platform::windows::autostart`.
 
 #[tauri::command]
 pub fn set_autostart(enabled: bool) -> Result<(), String> {
-    let path = desktop_entry_path();
-    if enabled {
-        let dir = autostart_dir();
-        std::fs::create_dir_all(&dir)
-            .map_err(|e| format!("Failed to create autostart dir: {e}"))?;
-        let exe = current_exe_path();
-        let contents = format!(
-            "[Desktop Entry]\n\
-             Type=Application\n\
-             Name=Look\n\
-             Exec={exe}\n\
-             Icon=look\n\
-             Comment=Desktop launcher\n\
-             X-GNOME-Autostart-enabled=true\n\
-             StartupNotify=false\n"
-        );
-        std::fs::write(&path, contents).map_err(|e| format!("Failed to write autostart file: {e}"))
-    } else {
-        if path.exists() {
-            std::fs::remove_file(&path).map_err(|e| format!("Failed to remove autostart file: {e}"))
-        } else {
-            Ok(())
-        }
+    #[cfg(target_os = "linux")]
+    {
+        crate::platform::linux::autostart::set(enabled)
+    }
+
+    #[cfg(target_os = "windows")]
+    {
+        crate::platform::windows::autostart::set(enabled)
+    }
+
+    #[cfg(not(any(target_os = "linux", target_os = "windows")))]
+    {
+        let _ = enabled;
+        Ok(())
     }
 }
 
 #[tauri::command]
 pub fn get_autostart() -> bool {
-    desktop_entry_path().exists()
+    #[cfg(target_os = "linux")]
+    {
+        crate::platform::linux::autostart::get()
+    }
+
+    #[cfg(target_os = "windows")]
+    {
+        crate::platform::windows::autostart::get()
+    }
+
+    #[cfg(not(any(target_os = "linux", target_os = "windows")))]
+    {
+        false
+    }
 }
